Pipeline and PipelineMonitor now refactored to be separate from GitLabMonitor
