+++ lib/gitlab_monitor/gitlab_monitor.ex
@@ -14,54 +14,10 @@ defmodule CncfDashboardApi.GitlabMonitor do
     CncfDashboardApi.GitlabMonitor.Dashboard.last_checked()
     migrate_source_key_monitor(source_key_project_monitor_id)
     |> upsert_pipeline_monitor_info
-    |> upsert_ref_monitor
+    |> upsert_gitlab_to_ref_monitor 
     CncfDashboardApi.GitlabMonitor.Dashboard.broadcast()
     :ok
   end 
-
-
- @doc """
-  Get the source key models for the source key monitor, the project, and the pipeline 
-  based on `source_key_project_monitor_id`.
-
-  Source keys map the keys in gitlab to the keys in our database.
-
-  A migration from gitlab must have occured before calling this function in order to get 
-  valid data 
-
-  Returns `{:ok, monitor, source_key_project, source_key_pipeline}`
-  """
-  def source_models(source_key_project_monitor_id) do
-    monitor = Repo.all(from skpm in CncfDashboardApi.SourceKeyProjectMonitor, 
-                                        where: skpm.id == ^source_key_project_monitor_id) |> List.first
-    source_key_project = Repo.all(from skp in CncfDashboardApi.SourceKeyProjects, 
-                                                   where: skp.source_id == ^monitor.source_project_id) |> List.first
-    source_key_pipeline = Repo.all(from skp in CncfDashboardApi.SourceKeyPipelines, 
-                                                   where: skp.source_id == ^monitor.source_pipeline_id) |> List.first
-    {:ok, monitor, source_key_project, source_key_pipeline}
-  end
-
- @doc """
-  Retrieve the pipeline monitor record based on the `source_key_project_monitor_id`.
-  The pipeline monitor record maintains the local keys for monitored pipelines and
-  is keyed based on project_id,  pipeline_id and release_type
-
-  Returns `{found, record}`.
-  """
-  def pipeline_monitor(source_key_project_monitor_id) do
-    {:ok, monitor, source_key_project, source_key_pipeline} = source_models(source_key_project_monitor_id)
-
-    case is_deploy_pipeline_type(source_key_project.new_id) do
-      true -> pipeline_type = "deploy"
-      _ -> pipeline_type = "build"
-    end
-
-    {pm_found, pm_record} = %CncfDashboardApi.PipelineMonitor{pipeline_id: source_key_pipeline.new_id, 
-      project_id: source_key_project.new_id,
-      pipeline_type: pipeline_type,
-      release_type: monitor.pipeline_release_type} 
-      |> find_by([:pipeline_id, :project_id, :pipeline_type, :release_type])
-  end
 
  @doc """
   Checks if the target project has been migrated based on `project_name` and `source_pipeline_id`.
@@ -161,7 +117,7 @@ defmodule CncfDashboardApi.GitlabMonitor do
     CncfDashboardApi.GitlabMonitor.Dashboard.last_checked()
 
     # determine pipeline type
-    case is_deploy_pipeline_type(source_key_project.new_id) do
+    case CncfDashboardApi.GitlabMonitor.Pipeline.is_deploy_pipeline_type(source_key_project.new_id) do
       true -> pipeline_type = "deploy"
       _ -> pipeline_type = "build"
     end
@@ -215,155 +171,46 @@ defmodule CncfDashboardApi.GitlabMonitor do
     CncfDashboardApi.GitlabMigrations.upsert_pipeline_jobs(monitor.source_project_id |> String.to_integer, 
                                                            monitor.source_pipeline_id |> String.to_integer)
 
-    # upsert_ref_monitor(source_key_project.new_id,source_key_pipeline.new_id)
-    
-    # CncfDashboardApi.GitlabMonitor.Dashboard.broadcast()
-
-    project_pipeline_info = project_pipeline_info(source_key_project.new_id, source_key_pipeline.new_id)
+    project_pipeline_info = CncfDashboardApi.GitlabMonitor.Pipeline.project_pipeline_info(source_key_project.new_id, source_key_pipeline.new_id)
   end
-
-  # project name is either cross-cloud (cross-cloud handles the deploy pipelines)
-  # or other (a build pipeline)
-  def is_deploy_pipeline_type(project_id) do
-    project = Repo.all(from skp in CncfDashboardApi.Projects, 
-                       where: skp.id == ^project_id) |> List.first
-    Logger.info fn ->
-      "is_deploy_pipeline_type project: #{inspect(project)}"
-    end
-    if (project.name =~ "cross-cloud" || project.name =~ "cross-project") do
-      true
-    else
-      false
-    end
-
-  end
-
- @doc """
- Gets all project and pipeline information based on the internal
- `project_id` and  `pipeline_id`.
-
-  project, pipeline, and pipeline jobs should be migrated before
-    calling project_pipeline_info 
-
-  Returns `{:ok, %Projects, %Pipelines, [%PipelineJobs], %PipelineMonitor]}`
-  """
-  def project_pipeline_info(project_id, pipeline_id) do
-    Logger.info fn ->
-      "project_pipeline_info project id: #{project_id} pipeline_id: #{pipeline_id}"
-    end
-    project = Repo.all(from p in CncfDashboardApi.Projects, 
-                       where: p.id == ^project_id) |> List.first
-                       # get pipeline
-    pipeline = Repo.all(from p in CncfDashboardApi.Pipelines, 
-                        where: p.id == ^pipeline_id) |> List.first
-                        # get pipeline jobs
-    pipeline_jobs = Repo.all(from pj in CncfDashboardApi.PipelineJobs, 
-                             where: pj.pipeline_id == ^pipeline_id)
-
-    {pm_found, pipeline_monitor} = %CncfDashboardApi.PipelineMonitor{pipeline_id: pipeline.id, 
-      project_id: project_id} |> find_by([:pipeline_id, :project_id])
-
-    Logger.info fn ->
-      "project_pipeline_info pipeline_monitor: #{inspect(pipeline_monitor)}"
-    end
-
-    {:ok, project, pipeline, pipeline_jobs, pipeline_monitor}
-  end
-
-
- @doc """
-  Gets the corresponding build pipeline based on `%PipelineMonitor`.
-  Returns `%PipelineMonitor`
-  """
-  def build_pipeline_monitor_by_deploy_pipeline_monitor(deploy_pipeline_monitor) do
-    if deploy_pipeline_monitor.pipeline_type == "deploy" do
-      Repo.all(from pm in CncfDashboardApi.PipelineMonitor, 
-                           where: pm.pipeline_id == ^deploy_pipeline_monitor.internal_build_pipeline_id, 
-                           where: pm.pipeline_type == "build") 
-                           |> List.first()
-    else
-      # pipeline is a build pipeline
-      deploy_pipeline_monitor
-    end
-  end
-
- @doc """
-  Selects the target pipeline info based on `{%PipelineMonitor, %Pipeline}`.
-
-  target pipeline is the current (passed in) pipeline and pipeline monitor
-  if the pipeline type is `deploy`
-
-  target pipeline is based on the internal_build_pipeline_id 
-  if the pipeline type is `build`
-
-  Returns `{%PipelineMonitor, %Pipelines}`
-  """
- def target_pipeline_info(pipeline_monitor, pipeline) do
-    if pipeline_monitor.pipeline_type == "deploy" do
-      target_pm = build_pipeline_monitor_by_deploy_pipeline_monitor(pipeline_monitor)
-      target_pl = Repo.all(from pm in CncfDashboardApi.Pipelines, 
-                           where: pm.id == ^pipeline_monitor.internal_build_pipeline_id ) |> List.first
-
-      if is_nil(target_pm) do
-        Logger.error fn ->
-          "Legacy dependency.  A deploy call with no build calls has been found for: #{inspect(pipeline_monitor)}"
-        end
-        raise "There should be no deploy pipelines that do not have a corresponding build pipeline 
-        Must be a legacy call.  This deploy pipeline call is probably for a dependency that no 
-        longer exists in the db"
-        # target_pm = pipeline_monitor
-        # target_pl = pipeline
-      end
-      Logger.info fn ->
-        " upsert_ref_monitor deploy target_pm: #{inspect(target_pm)}"
-      end
-    else
-      Logger.info fn ->
-        " upsert_ref_monitor build target_pm: #{inspect(pipeline_monitor)}"
-      end
-      target_pm = pipeline_monitor
-      target_pl = pipeline
-    end
-    {target_pm, target_pl}
- end
 
  @doc """
   Updates one ref monitor based on `project_id` and  `pipeline_id`.
 
   project, pipeline, and pipeline jobs should be migrated before
-    calling upsert_ref_monitor
+    calling upsert_gitlab_to_ref_monitor 
 
   1. source_key_project_monitor is called from a http post
   2. pipeline_monitor is created/updated during the http_post
-  3. upsert_ref_monitor is called to set up the dashboard
+  3. upsert_gitlab_to_ref_monitor is called to set up the dashboard
 
   Returns `[%DashboardBadgeStatus]`
   """
-  def upsert_ref_monitor(project_id, pipeline_id) do
+  def upsert_gitlab_to_ref_monitor(project_id, pipeline_id) do
     Logger.info fn ->
       "upsert_ref_monitor project id: #{project_id} pipeline_id: #{pipeline_id}"
     end
-    project_pipeline_info = project_pipeline_info(project_id, pipeline_id)
-    upsert_ref_monitor(project_pipeline_info)
+    project_pipeline_info = CncfDashboardApi.GitlabMonitor.Pipeline.project_pipeline_info(project_id, pipeline_id)
+    upsert_gitlab_to_ref_monitor(project_pipeline_info)
   end
 
-  def upsert_ref_monitor({:ok, project, pipeline, pipeline_jobs, pipeline_monitor} = project_pipeline_monitor) do
-    upsert_ref_monitor({project, pipeline, pipeline_jobs, pipeline_monitor})
+  def upsert_gitlab_to_ref_monitor({:ok, project, pipeline, pipeline_jobs, pipeline_monitor} = project_pipeline_monitor) do
+    upsert_gitlab_to_ref_monitor({project, pipeline, pipeline_jobs, pipeline_monitor})
   end
 
  @doc """
   Updates one ref monitor based on `{%Project, %Pipeline, [%PipelineJobs], %PipelineMonitor}`.
 
   project, pipeline, and pipeline jobs should be migrated before
-    calling upsert_ref_monitor
+    calling upsert_gitlab_to_ref_monitor 
 
   1. source_key_project_monitor is called from a http post
   2. pipeline_monitor is created/updated during the http_post
-  3. upsert_ref_monitor is called to set up the dashboard
+  3. upsert_gitlab_to_ref_monitor is called to set up the dashboard
 
   Returns `[%DashboardBadgeStatus]`
   """
-  def upsert_ref_monitor({project, pipeline, pipeline_jobs, pipeline_monitor} = project_pipeline_monitor) do
+  def upsert_gitlab_to_ref_monitor({project, pipeline, pipeline_jobs, pipeline_monitor} = project_pipeline_monitor) do
 
     Logger.info fn ->
       "upsert_ref_monitor/1 project id: #{project.id} pipeline_id: #{pipeline.id}"
@@ -396,7 +243,7 @@ defmodule CncfDashboardApi.GitlabMonitor do
       " upsert_ref_monitor all build pipeline_monitors: #{inspect(Repo.all(CncfDashboardApi.PipelineMonitor) )}"
     end
 
-    {target_pm, target_pl} = target_pipeline_info(pipeline_monitor, pipeline)
+    {target_pm, target_pl} = CncfDashboardApi.GitlabMonitor.Pipeline.target_pipeline_info(pipeline_monitor, pipeline)
 
     rm_record = CncfDashboardApi.GitlabMonitor.Dashboard.upsert_ref_monitor(pipeline_monitor, target_pm, target_pl, pipeline_order)
