+++ test/lib/gitlab_monitor.pipeline_monitor_test.exs
@@ -1,0 +1,14 @@
+require IEx;
+# require CncfDashboardApi.DataMigrations;
+require Logger;
+defmodule CncfDashboardApi.GitlabMonitor.PipelineMonitorTest do
+  use CncfDashboardApi.ChannelCase
+
+  alias CncfDashboardApi.DashboardChannel
+
+  import Ecto.Query
+  import CncfDashboardApi.Factory
+  use ExUnit.Case
+  
+
+end
